Merging the change of type for all the ids

See merge request SPEAM/2024-2025/team3/mdtools!71

+++ backend/src/main/java/be/uliege/speam/team03/MDTools/compositeKeys/AlternativesKey.java
@@ -17,10 +17,10 @@ import lombok.Setter;
 @Embeddable
 public class AlternativesKey implements Serializable{
     @Column(name = "instruments_id_1")
-    private Integer instrId1;
+    private Long instrId1;
 
     @Column(name = "instruments_id_2")
-    private Integer instrId2;
+    private Long instrId2;
 
     @Override
     public boolean equals(Object o){
